requirements.txt
# ============================================================
# AI Training Platform - Consolidated Dependencies
# ============================================================
# Shared requirements for all courses and modules.
# Install from any module folder: pip install -r <root>/requirements.txt

# --- Core ---
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0

# --- LLM Providers ---
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99

# --- Agents ---
langgraph>=0.1.0
numexpr>=2.8.0

# --- Document Processing ---
pypdf>=3.0.0
unstructured>=0.10.0
unstructured[markdown]>=0.10.0
markdown>=3.4.0
wikipedia>=1.4.0

# --- Vector Stores ---
langchain-chroma>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# --- Advanced ---
langchain-experimental>=0.0.50
tiktoken>=0.5.0
rank-bm25>=0.2.2

# --- Graph Databases ---
langchain-neo4j>=0.1.0
neo4j>=5.0.0

# --- Evaluation ---
ragas>=0.1.0
langsmith>=0.1.0

# --- Interactive Learning ---
rich>=13.0.0
InquirerPy>=0.3.4
rich-pixels>=3.0.0
